pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openaiagents"
version = "0.1.0"
description = "Python + DevContainer LLM Agent learning sandbox (FastAPI, PPTX, workflow DAG)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "You" }]

dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.27,<1.0",
  "pydantic>=2.6,<3.0",
  "python-pptx>=0.6.23,<1.0",
  "Pillow>=10.2,<12.0",
  "httpx>=0.27,<1.0",
  "python-dotenv>=1.0,<2.0",
  "networkx>=3.2,<4.0",
  "rich>=13.7,<14.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.6,<0.9",
  "mypy>=1.8,<2.0",
  "pre-commit>=3.6,<5.0",
  "types-Pillow>=10.2.0.20240520",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true


